This commit was manufactured by cvs2svn to create tag '1.8.12'.
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/tags/1.8.12@11512 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/calculation/fin.c
@@ -2418,7 +2418,7 @@ Amortization_Schedule (amort_sched_ptr amortsched)
 
     for (jj = n, j = 0; jj > 0; jj -= yr_pmt, yr_pmt = PF, j++)
     {
-      if (jj <= PF)
+      if (jj <= (int)PF)
       {
 	yr_fv = fv;
 	yr_int = rnd (((jj - 1) * pmt) + hpv + final_pmt, prec);
