+++ src/engine/gnc-engine-util.c
@@ -372,7 +372,7 @@ ultostr (unsigned long val, int base)
   }
 
   /* print */
-  for (i=0; i<places; i++) {
+  for (i=0; i<(int)places; i++) {
     if (10>broke[i]) {
        buf[places-1-i] = 0x30+broke[i];  /* ascii digit zero */
     } else {
